chore: move input component
