+++ apps/web/src/components/Composer/Actions/CollectSettings/AmountConfig.tsx
@@ -1,10 +1,10 @@
 import ToggleWithHelper from '@components/Shared/ToggleWithHelper';
-import { Input } from '@components/UI/Input';
 import { CollectionIcon } from '@heroicons/react/outline';
 import { t, Trans } from '@lingui/macro';
 import type { Erc20 } from 'lens';
 import type { FC } from 'react';
 import { useCollectModuleStore } from 'src/store/collect-module';
+import { Input } from 'ui';
 
 import ReferralConfig from './ReferralConfig';
