+++ apps/web/src/components/Composer/Actions/CollectSettings/SplitConfig.tsx
@@ -1,6 +1,5 @@
 import Beta from '@components/Shared/Badges/Beta';
 import ToggleWithHelper from '@components/Shared/ToggleWithHelper';
-import { Input } from '@components/UI/Input';
 import { PlusIcon, SwitchHorizontalIcon, UsersIcon, XCircleIcon } from '@heroicons/react/outline';
 import isValidEthAddress from '@lib/isValidEthAddress';
 import { t, Trans } from '@lingui/macro';
@@ -9,7 +8,7 @@ import { useProfileLazyQuery } from 'lens';
 import type { FC } from 'react';
 import { useAppStore } from 'src/store/app';
 import { useCollectModuleStore } from 'src/store/collect-module';
-import { Button } from 'ui';
+import { Button, Input } from 'ui';
 import splitNumber from 'utils/splitNumber';
 
 interface SplitConfigProps {
