+++ apps/web/src/components/Composer/Actions/Giphy/GifSelector.tsx
@@ -1,4 +1,3 @@
-import { Input } from '@components/UI/Input';
 import { useDebounce } from '@components/utils/hooks/useDebounce';
 import type { ICategory } from '@giphy/js-fetch-api';
 import { GiphyFetch } from '@giphy/js-fetch-api';
@@ -7,6 +6,7 @@ import { Grid } from '@giphy/react-components';
 import { t, Trans } from '@lingui/macro';
 import type { ChangeEvent, Dispatch, FC } from 'react';
 import { useEffect, useState } from 'react';
+import { Input } from 'ui';
 
 const giphyFetch = new GiphyFetch('sXpGFDGZs0Dv1mmNFvYaGUvYwKX0PWIh');
