+++ apps/web/src/components/Messages/Composer.tsx
@@ -1,4 +1,3 @@
-import { Input } from '@components/UI/Input';
 import { Spinner } from '@components/UI/Spinner';
 import useWindowSize from '@components/utils/hooks/useWindowSize';
 import { ArrowRightIcon } from '@heroicons/react/outline';
@@ -10,7 +9,7 @@ import { useEffect, useState } from 'react';
 import toast from 'react-hot-toast';
 import { useMessagePersistStore } from 'src/store/message';
 import { MESSAGES } from 'src/tracking';
-import { Button } from 'ui';
+import { Button, Input } from 'ui';
 
 interface ComposerProps {
   sendMessage: (message: string) => Promise<boolean>;
