+++ apps/web/src/components/Settings/Profile/NftPicture.tsx
@@ -1,4 +1,3 @@
-import { Input } from '@components/UI/Input';
 import { Spinner } from '@components/UI/Spinner';
 import { PencilIcon } from '@heroicons/react/outline';
 import { Mixpanel } from '@lib/mixpanel';
@@ -19,7 +18,7 @@ import { useState } from 'react';
 import toast from 'react-hot-toast';
 import { useAppStore } from 'src/store/app';
 import { SETTINGS } from 'src/tracking';
-import { Button, ErrorMessage, Form, useZodForm } from 'ui';
+import { Button, ErrorMessage, Form, Input, useZodForm } from 'ui';
 import getSignature from 'utils/getSignature';
 import { useContractWrite, useSignMessage, useSignTypedData } from 'wagmi';
 import { mainnet, polygon, polygonMumbai } from 'wagmi/chains';
