+++ apps/web/src/components/Shared/Login/New/index.tsx
@@ -1,5 +1,4 @@
 import ChooseFile from '@components/Shared/ChooseFile';
-import { Input } from '@components/UI/Input';
 import { Spinner } from '@components/UI/Spinner';
 import { PlusIcon } from '@heroicons/react/outline';
 import uploadToIPFS from '@lib/uploadToIPFS';
@@ -8,7 +7,7 @@ import { APP_NAME, HANDLE_REGEX, ZERO_ADDRESS } from 'data/constants';
 import { useCreateProfileMutation } from 'lens';
 import type { ChangeEvent, FC } from 'react';
 import { useState } from 'react';
-import { Button, ErrorMessage, Form, useZodForm } from 'ui';
+import { Button, ErrorMessage, Form, Input, useZodForm } from 'ui';
 import getStampFyiURL from 'utils/getStampFyiURL';
 import { useAccount } from 'wagmi';
 import { object, string } from 'zod';
