+++ apps/web/src/components/Shared/Navbar/Search.tsx
@@ -1,4 +1,3 @@
-import { Input } from '@components/UI/Input';
 import { Spinner } from '@components/UI/Spinner';
 import useOnClickOutside from '@components/utils/hooks/useOnClickOutside';
 import { SearchIcon, XIcon } from '@heroicons/react/outline';
@@ -9,7 +8,7 @@ import { CustomFiltersTypes, SearchRequestTypes, useSearchProfilesLazyQuery } fr
 import { useRouter } from 'next/router';
 import type { ChangeEvent, FC } from 'react';
 import { useRef, useState } from 'react';
-import { Card } from 'ui';
+import { Card, Input } from 'ui';
 import formatHandle from 'utils/formatHandle';
 
 import UserProfile from '../UserProfile';
